requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-chroma==0.2.0
langchain-ollama==0.2.0
pydantic==2.10.3
pydantic-settings==2.4.0
pypdf==5.1.0
python-docx==1.1.2
streamlit==1.40.2
chromadb==0.5.21
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
requests==2.32.3
colorlog==6.8.2
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Optional LLM Providers (install as needed)
langchain-openai==0.1.20  # For OpenAI API
langchain-groq==0.1.3     # For Groq API
langchain-huggingface==0.0.16  # For local HuggingFace models

